✨ feat : Member 요청 Dto 및 Controller 구현 (Enum 부분 수정 필요)
