+++ BE/UYouBooDan/src/main/java/TeamBigDipper/UYouBooDan/member/controller/MemberController.java
@@ -2,7 +2,8 @@ package TeamBigDipper.UYouBooDan.member.controller;
 
 import TeamBigDipper.UYouBooDan.global.dto.MultiResDto;
 import TeamBigDipper.UYouBooDan.global.dto.SingleResDto;
-import TeamBigDipper.UYouBooDan.member.dto.MemberReqDto;
+import TeamBigDipper.UYouBooDan.member.dto.MemberPatchReqDto;
+import TeamBigDipper.UYouBooDan.member.dto.MemberPostReqDto;
 import TeamBigDipper.UYouBooDan.member.entity.Member;
 import TeamBigDipper.UYouBooDan.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
@@ -28,21 +29,30 @@ public class MemberController {
      * @return "data" : "String"
      */
     @PostMapping
-    public ResponseEntity<SingleResDto<Long>> postMember (@RequestBody MemberReqDto memberReqDto) {
-        Member savedMember = memberService.createMember(memberReqDto.toMember());
+    public ResponseEntity<SingleResDto<Long>> postMember (@RequestBody MemberPostReqDto memberPostReqDto) {
+        /**
+         * 추후 시큐리티 구현 후 member password 암호화 예정
+         */
+        Member savedMember = memberService.createMember(memberPostReqDto.toMember());
 
-        return new ResponseEntity<>(new SingleResDto<>(savedMember.getMember_id()), HttpStatus.CREATED);
+        return new ResponseEntity<>(new SingleResDto<>(savedMember.getMemberId()), HttpStatus.CREATED);
     }
+
 
     /**
      * Patch 요청
      * @return "data" : "String"
+     * @param memberPatchReqDto : 요청 Body
+     * @param memberId : 시큐리티 구성 전 임시 Variable
+     * @return void
      */
-    @PatchMapping("/edit")
-    public ResponseEntity<SingleResDto<String>> patchMember () {
-
+    @PatchMapping("/edit/{member_id}")
+    public ResponseEntity<SingleResDto<String>> patchMember (@RequestBody MemberPatchReqDto memberPatchReqDto,
+                                                             @PathVariable("member_id") Long memberId) {
+        memberService.modifyMember(memberPatchReqDto.toMember(), memberId);
         return new ResponseEntity<>(new SingleResDto<>("Success Patch"), HttpStatus.OK);
     }
+
 
     /**
      * Delete 요청
@@ -52,9 +62,12 @@ public class MemberController {
      */
     @DeleteMapping("/remove")
     public ResponseEntity<SingleResDto<String>> deleteMember () {
-        // 로그인한 사용자 Id값을 받아 상태 변환 예정
+        /**
+         * 로그인한 사용자 Id값을 받아 상태 변환 예정
+         */
         return new ResponseEntity<>(new SingleResDto<>("Success Delete"), HttpStatus.OK);
     }
+
 
     /**
      * 단일 Get 요청
@@ -66,6 +79,7 @@ public class MemberController {
         return new ResponseEntity<>(new SingleResDto<>("Success Get"), HttpStatus.OK);
     }
 
+
     /**
      * Page Get 요청
      * @return "data" : "String"
@@ -73,8 +87,8 @@ public class MemberController {
     @GetMapping("/find-all")
     public ResponseEntity<MultiResDto<String>> getMembers (Pageable pageable) {
         List<String> list = new ArrayList<>();
-        int count = 1;
-        while(count++<=30) list.add("Member " + count);
+        int count = 0;
+        while(count++<30) list.add("Member " + count);
         Page<String> page = new PageImpl<>(list, pageable, list.size());
 
         return new ResponseEntity<>(new MultiResDto<>(list, page), HttpStatus.OK);
